@@ -7,15 +7,16 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
+	"net"
+	"path/filepath"
+	"testing"
+	"time"
+
 	"github.com/go-logr/logr"
 	fuzz "github.com/google/gofuzz"
 	"github.com/smart-edge-open/sriov-fec-operator/pkg/common/utils"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
-	"net"
-	"path/filepath"
-	"testing"
-	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
