@@ -33,23 +33,6 @@ func (in *ACC100BBDevConfig) DeepCopy() *ACC100BBDevConfig {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ACC200BBDevConfig) DeepCopyInto(out *ACC200BBDevConfig) {
-	*out = *in
-	out.ACC100BBDevConfig = in.ACC100BBDevConfig
-	out.QFFT = in.QFFT
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ACC200BBDevConfig.
-func (in *ACC200BBDevConfig) DeepCopy() *ACC200BBDevConfig {
-	if in == nil {
-		return nil
-	}
-	out := new(ACC200BBDevConfig)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *AcceleratorSelector) DeepCopyInto(out *AcceleratorSelector) {
 	*out = *in
 }
@@ -465,6 +448,23 @@ func (in *VF) DeepCopy() *VF {
 		return nil
 	}
 	out := new(VF)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ACC200BBDevConfig) DeepCopyInto(out *ACC200BBDevConfig) {
+	*out = *in
+	out.ACC100BBDevConfig = in.ACC100BBDevConfig
+	out.QFFT = in.QFFT
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ACC200BBDevConfig.
+func (in *ACC200BBDevConfig) DeepCopy() *ACC200BBDevConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(ACC200BBDevConfig)
 	in.DeepCopyInto(out)
 	return out
 }
